update: parse: expansion
