+++ srcs/parse/expand_heredoc.c
@@ -1,0 +1,2 @@
+#include "minishell.h"
+
